Add arm64 support to VM bookkeeping library
