+++ Sources/libhostmgr/Model/Paths.swift
@@ -2,6 +2,8 @@ import Foundation
 
 public struct Paths {
 
+    private static let storageDirectoryIdentifier = "com.automattic.hostmgr"
+
     private static var homebrewRoot: URL {
         switch ProcessInfo.processInfo.processorArchitecture {
         case .arm64: return URL(fileURLWithPath: "/opt/homebrew", isDirectory: true)
@@ -10,19 +12,40 @@ public struct Paths {
     }
 
     static var storageRoot: URL {
+        #if arch(arm64)
+        arm64StorageRoot.appendingPathComponent(storageDirectoryIdentifier, isDirectory: true)
+        #else
         homebrewRoot.appendingPathComponent("var", isDirectory: true)
+        #endif
+    }
+
+    /// A way to push the available check out to keep things pretty
+    private static var arm64StorageRoot: URL {
+        guard #available(macOS 13.0, *) else {
+            preconditionFailure("Apple Silicon in CI should only run on macOS 13 or greater")
+        }
+
+        return URL.applicationSupportDirectory
     }
 
     static var configurationRoot: URL {
-        homebrewRoot
+        #if arch(arm64)
+        storageRoot.appendingPathComponent("configuration")
+        #else
+        storageRoot
             .appendingPathComponent("etc", isDirectory: true)
             .appendingPathComponent("hostmgr", isDirectory: true)
+        #endif
     }
 
     static var stateRoot: URL {
+        #if arch(arm64)
+        storageRoot.appendingPathComponent("state")
+        #else
         storageRoot
             .appendingPathComponent("hostmgr", isDirectory: true)
             .appendingPathComponent("state", isDirectory: true)
+        #endif
     }
 
     public static var vmImageStorageDirectory: URL {
