+++ Sources/libhostmgr/libhostmgr.swift
@@ -1,5 +1,4 @@
 import Foundation
-import prlctl
 
 /// Downloads and registers an image by name
 ///
@@ -19,8 +18,9 @@ public func downloadRemoteImage(
     remoteRepository: RemoteVMRepository? = nil
 ) async throws -> URL {
     let remoteRepository = try remoteRepository ?? RemoteVMRepository()
-    guard let remoteImage = try await remoteRepository.getImage(named: name) else {
-        Console.crash(message: "Unable to find remote image: \(name)", reason: .unableToFindRemoteImage)
+
+    guard let remoteImage = try await remoteRepository.getCompatibleImage(named: name) else {
+        Console.crash(message: "Unable to find an image named `\(name)` for \(ProcessInfo.processInfo.processorArchitecture)", reason: .unableToFindRemoteImage)
     }
 
     return try await downloadRemoteImage(remoteImage)
@@ -75,31 +75,14 @@ public func unpackVM(name: String) async throws {
         Console.crash(message: "Local VM \(name) could not be found", reason: .fileNotFound)
     }
 
-    guard let importedVirtualMachine = try Parallels().importVM(at: localVM.path) else {
-        Console.crash(message: "Unable to import VM at: \(localVM.path)", reason: .unableToImportVM)
+    #if arch(arm64)
+    guard #available(macOS 13.0, *) else {
+        preconditionFailure("Apple Silicon in CI should only run on macOS 13 or greater")
     }
-
-    guard localVM.state == .packaged, let package = importedVirtualMachine.asPackagedVM() else {
-        Console.crash(message: "VM \(name) is not a packaged VM", reason: .invalidVMStatus)
-    }
-
-    Console.info("Unpacking \(package.name) – this will take a few minutes")
-    let unpackedVM = try package.unpack()
-    Console.success("Unpacked \(package.name)")
-
-    // If we simply rename the `.pvmp` file, the underlying `pvm` file may retain its original name. We should
-    // update the file on disk to reference this name
-    if name != unpackedVM.name {
-        Console.info("Fixing Parallels VM Label")
-        try unpackedVM.rename(to: name)
-        Console.success("Parallels VM Label Fixed")
-    }
-
-    Console.success("Finished Unpacking VM")
-
-    Console.info("Cleaning Up")
-    try unpackedVM.unregister()
-    Console.success("Done")
+    try Compressor().decompress(archiveAt: localVM.path, to: Paths.toAppleSiliconVM(named: name))
+    #else
+    try await ParallelsVMRepository().unpack(localVM: localVM)
+    #endif
 }
 
 /// Resets local VM storage by removing all registered Parallels VMs and temporary VM clones.
@@ -151,7 +134,7 @@ public func listAvailableRemoteImages(
 ) async throws -> [RemoteVMImage] {
     let remoteRepository = try remoteRepository ?? RemoteVMRepository()
     let manifest = try await remoteRepository.getManifest()
-    let remoteImages = try await remoteRepository.listImages(sortedBy: strategy)
+    let remoteImages = try await remoteRepository.listCompatibleImages(sortedBy: strategy)
     let localImages = try localRepository.list()
 
     return remoteImages
@@ -175,84 +158,27 @@ public func listLocalImagesToDelete(
 }
 
 // MARK: Virtual Machine Control
-public func lookupParallelsVMOrExit(
-    withIdentifier id: String,
-    parallelsRepository: ParallelsVMRepository = ParallelsVMRepository()
-) throws -> VM {
-    guard let parallelsVirtualMachine = try parallelsRepository.lookupVM(byIdentifier: id) else {
-        Console.crash(
-            message: "There is no VM with the name or UUID `\(id)` registered with Parallels",
-            reason: .parallelsVirtualMachineDoesNotExist
-        )
-    }
-
-    return parallelsVirtualMachine
-}
-
 public func startVM(name: String) async throws {
+    #if arch(arm64)
     try await XPCService.startVM(named: "test")
+    #else
+    try await ParallelsVMRepository().startVM(named: name)
+    #endif
 
-    return
-
-    let startDate = Date()
-
-    guard let sourceVM = try LocalVMRepository().lookupVM(withName: name) else {
-        Console.crash(message: "VM \(name) could not be found", reason: .fileNotFound)
-    }
-
-    try resetVMStorage()
-
-    if sourceVM.state == .packaged {
-        try await unpackVM(name: name)
-        return try await startVM(name: name)
-    }
-
-    let destination = FileManager.default.temporaryFilePath(named: name + ".tmp.pvm")
-    try FileManager.default.removeItemIfExists(at: destination)
-    try FileManager.default.copyItem(at: sourceVM.path, to: destination)
-    Console.info("Created temporary VM at \(destination)")
-
-    guard let parallelsVM = try Parallels().importVM(at: destination)?.asStoppedVM() else {
-        Console.crash(message: "Unable to import VM: \(destination)", reason: .unableToImportVM)
-    }
-
-    Console.success("Successfully Imported \(parallelsVM.name) with UUID \(parallelsVM.uuid)")
-
-    try applyVMSettings([
-        .memorySize(Int(ProcessInfo().physicalMemory / 1024 / 1024) - 4096),  // We should make this configurable
-        .cpuCount(ProcessInfo().physicalProcessorCount),
-        .hypervisorType(.apple),
-        .networkType(.shared),
-        .isolateVM(.on),
-        .sharedCamera(.off)
-    ], to: parallelsVM)
-
-    try parallelsVM.start(wait: false)
-
-    let _: Void = try await withCheckedThrowingContinuation { continuation in
-        do {
-            try waitForVMStartup(parallelsVM)
-        } catch {
-            continuation.resume(with: .failure(error))
-            return
-        }
-
-        let elapsed = Date().timeIntervalSince(startDate)
-        Console.success("Booted \(parallelsVM.name) \(Format.time(elapsed))")
-        continuation.resume(with: .success(()))
-    }
+    Console.success("VM is running")
 }
 
 public func stopAllRunningVMs(
     immediately: Bool = true,
     parallelsRepository: ParallelsVMRepositoryProtocol = ParallelsVMRepository()
 ) async throws {
+    #if arch(arm64)
     try await XPCService.stopVM()
-    return
+    #else
+    ParallelsVMRepository().stopAllRunningVMs(immediately: immediately)
+    #endif
 
-    for parallelsVM in try parallelsRepository.lookupRunningVMs() {
-        try await stopRunningVM(name: parallelsVM.name, immediately: immediately)
-    }
+    Console.success("Shutdown Complete")
 }
 
 public func stopRunningVM(
@@ -260,70 +186,11 @@ public func stopRunningVM(
     immediately: Bool,
     parallelsRepository: ParallelsVMRepositoryProtocol = ParallelsVMRepository()
 ) async throws {
+    #if arch(arm64)
     try await XPCService.stopVM()
-    return
-
-    let parallelsVM = try lookupParallelsVMOrExit(withIdentifier: name)
-
-    guard let vmToStop = parallelsVM.asRunningVM() else {
-        Console.exit(
-            message: "\(parallelsVM.name) is not running, so it can't be stopped",
-            style: .warning
-        )
-    }
-
-    Console.info("Shutting down \(parallelsVM.name)")
-
-    try vmToStop.shutdown(immediately: immediately)
-    try vmToStop.unregister()
-
-    // Clean up after ourselves by deleting the VM files
-    let vmPath = FileManager.default.temporaryDirectory.appendingPathComponent(parallelsVM.name + ".tmp.pvm")
-
-    guard try FileManager.default.directoryExists(at: vmPath) else {
-        Console.success("Shutdown Complete")
-        return
-    }
-
-    Console.info("Deleting VM storage from \(vmPath)")
-    try FileManager.default.removeItem(at: vmPath)
+    #else
+    ParallelsVMRepository().stopRunningVM(named: name, immediately: immediately)
+    #endif
 
     Console.success("Shutdown Complete")
-}
-
-func waitForVMStartup(_ parallelsVirtualMachine: StoppedVM) throws {
-    repeat {
-        usleep(100_000)
-    } while try Parallels()
-        .lookupRunningVMs()
-        .filter { $0.uuid == parallelsVirtualMachine.uuid && $0.hasIpV4Address }
-        .isEmpty
-}
-
-func applyVMSettings(_ settings: [StoppedVM.VMOption], to parallelsVM: StoppedVM) throws {
-    Console.info("Applying VM Settings")
-
-    // Always leave 4GB available to the VM host – the VM can have the rest
-    let dedicatedMemoryForVM = ProcessInfo().physicalMemory - (4096 * 1024 * 1024) // We should make this configurable
-    let cpuCoreCount = ProcessInfo().physicalProcessorCount
-
-    Console.printTable(data: [
-        ["Total System Memory", Format.memoryBytes(ProcessInfo().physicalMemory)],
-        ["VM System Memory", Format.memoryBytes(dedicatedMemoryForVM)],
-        ["VM CPU Cores", "\(cpuCoreCount)"],
-        ["Hypervisor Type", "apple"],
-        ["Networking Type", "bridged"]
-    ])
-
-    for setting in settings {
-        try parallelsVM.set(setting)
-    }
-
-    // These are optional, and it's possible they've already been removed, so they may fail
-    do {
-        try parallelsVM.set(.withoutSoundDevice())
-        try parallelsVM.set(.withoutCDROMDevice())
-    } catch {
-        Console.warn("Unable to remove device: \(error.localizedDescription)")
-    }
 }
